rename package dir to ata_db_models
